backend/spin/requirements.txt
psycopg[binary]==3.1.8
PyJWT==2.8.0